lua: Add new timer events in a new list instead of directly
Previously, we would add new events to the same list we are currently
iterating. This could lead to infinite loops because we could enact the
timer we just added.

Instead always add to a new list and merge the newlist into the current
list at the end of the frame to ensure that newly added callbacks will
be run the next frame at the earliest.